refactor: move fields to folder
